refactor: AuthenticationManager, SynchronizationManager, TokenManager 인터페이스 Domain으로 이동
