+++ Drinkin/Utils/SynchronizationManager/DefaultSynchronizationManager.swift
@@ -1,5 +1,5 @@
 //
-//  SynchronizationManager.swift
+//  DefaultSynchronizationManager.swift
 //  Drinkin
 //
 //  Created by kimseongjun on 2023/12/15.
@@ -7,12 +7,6 @@
 
 import Foundation
 import Combine
-
-protocol SynchronizationManager {
-    func isDataChanged()
-    func initialize()
-    func dataChangeStatusPublisher() -> AnyPublisher<Bool, Never>
-}
 
 final class DefaultSynchronizationManager: SynchronizationManager {
     private let isDataChangedSubject = CurrentValueSubject<Bool, Never>(false)
